+++ MATLAB/main/Q1_m2_KF_simple.m
@@ -1,14 +1,21 @@
 clc; clear; close all;
 
 %% declare
+% simulation
 Fs = 500;   % sampling freq [Hz]
 dt = 1/Fs;  % sampling period [sec]
+
+% plant
+parm.a = 4;  % viscous friction [N-m / rad/s]
+parm.b = 35; % gain [N-m/volt]
+parm.d_Coulomb_coeff = 5;       % Coulomb friction coeff [volt]
+parm.d_Coulomb_threshold = 3;   % Coulomb friction thrshould [rad/s]
 
 %% ODE
 % simulation
 time = 0:dt:1.5;
 IC = [0 0]';
-equ = @(t,IC)equ_plant(t,IC);
+equ = @(t,IC)plant_DC_motor(t,IC,parm);
 [t_sim,X_sim] = ode45(equ, time, IC);
 
 % extract data