+++ MATLAB/test/pendulum/m1_diff_with_LPF.m
@@ -5,17 +5,17 @@ clc; clear; close all;
 Fs = 500;   % sampling freq [Hz]
 dt = 1/Fs;  % sampling period [sec]
 
-% plant
-parm.a = 4;  % viscous friction [N-m / rad/s]
-parm.b = 35; % gain [N-m/volt]
-parm.d_Coulomb_coeff = 5;       % Coulomb friction coeff [volt]
-parm.d_Coulomb_threshold = 3;   % Coulomb friction thrshould [rad/s]
+% plant parameter
+plant_parm.a = 0.5;  % viscous friction [N-m / rad/s]
+plant_parm.b = 5;    % gain [N-m/volt]
+plant_parm.g = 9.81; % gravity
+plant_parm.l = 0.1;  % length [m]
 
 %% ODE
 % simulation
 time = 0:dt:1.5;
 IC = [0 0]';
-equ = @(t,IC)plant_DC_motor(t,IC,parm);
+equ = @(t,IC)plant_pendulum(t,IC,plant_parm);
 [t_sim,X_sim] = ode45(equ, time, IC);
 
 % extract data