+++ Change Log.md
@@ -1,10 +1,23 @@
 Change Log
 ======================
 
+[v0.0.2] 2023 0226
+---------------------------
+# added
+- pendulum system
+- pendulum system, diff and LPF
+- pendulum system, KF
+- pendulum system, EKF
+- pendulum system, EKF with general disturbance state
+- DC motor system, EKF with general disturbance state
+
+# changed
+- rename and move DC motor system
+
 
 [v0.0.1] 2023 0225
 ---------------------------
 # added
-reddit_Kalman_filter_question project
+- reddit_Kalman_filter_question project